+++ EcoLessonAPI/DTOs/EmpresaDTO.cs
@@ -2,24 +2,27 @@ namespace EcoLessonAPI.DTOs;
 
 public class EmpresaDTO
 {
-    public decimal IdEmpresa { get; set; }
+    // -----------------------------------------------------------------
+    // MUDANÇA 1: Trocado 'decimal' por 'long' (para o ID)
+    // -----------------------------------------------------------------
+    public long IdEmpresa { get; set; } // <-- MUDADO DE 'decimal'
+    
     public string RazaoSocial { get; set; } = string.Empty;
-    public string Cnpj { get; set; } = string.Empty;
-    public string EmailEmpresa { get; set; } = string.Empty;
-    public List<LinkDTO>? Links { get; set; }
+    public string Cnpj { get; set; } = string.Empty;
+    public string EmailEmpresa { get; set; } = string.Empty;
+    public List<LinkDTO>? Links { get; set; }
 }
 
 public class EmpresaCreateDTO
 {
-    public string RazaoSocial { get; set; } = string.Empty;
-    public string Cnpj { get; set; } = string.Empty;
-    public string EmailEmpresa { get; set; } = string.Empty;
+    public string RazaoSocial { get; set; } = string.Empty; // Correto (é texto)
+    public string Cnpj { get; set; } = string.Empty; // Correto (é texto)
+    public string EmailEmpresa { get; set; } = string.Empty; // Correto (é texto)
 }
 
 public class EmpresaUpdateDTO
 {
-    public string RazaoSocial { get; set; } = string.Empty;
-    public string Cnpj { get; set; } = string.Empty;
-    public string EmailEmpresa { get; set; } = string.Empty;
+    public string RazaoSocial { get; set; } = string.Empty; // Correto (é texto)
+    public string Cnpj { get; set; } = string.Empty; // Correto (é texto)
+    public string EmailEmpresa { get; set; } = string.Empty; // Correto (é texto)
 }
-
